Use cuc_timestamp as index field when reading old
Due to the IndexPager using the index field over the timestamp
field when paging results, the index field cannot be aliased if
paging of the results is to be possible. While this is something
that should be fixed in core, the easiest solution to the problem
in CheckUser is not using the aliased field until it has to be
used (i.e. when reading new).

QA steps:
* Change wgCheckUserMaximumRowCount to a small value for testing
 to allow paging to occur
* Load Special:CheckUser 'get edits' and 'get users'
* Page the results using the paging links
* Ensure no exception is thrown.

Bug: T338287
Change-Id: Ibc3ab6e3807b191e843313737d77cca698ea3ff4
(cherry picked from commit a4da47decffdacd3023a847b1c77836833958d82)