Move ::onLoadExtensionSchemaUpdates to hook handler file
Move the definition of ::onLoadExtensionSchemaUpdates to a
separate hook handler file to reduce the overall size of Hooks.php.
The code for ::onLoadExtensionSchemaUpdates is left unmodified,
except for the first line to allow the schema folder to be found
with the new location for the code.

This is the first in many patches that attempt to move several
methods in Hooks.php to separate files to decrease file size and
therefore increase maintainability.

QA steps:
* Ensure CI passes (which runs update.php for a first install of CU)
* Run update.php on a wiki that had CheckUser already installed and
ensure it runs without problems.

Bug: T328334
Change-Id: Ie5771f363e6cab115c6c266e48c42a458518d885