Move toollinks related hooks to a hook handler file
Move the code defined for the ContributionsToolLinksHook and
UserToolLinksEditHook hooks to a separate hook handler file for
hooks related to tool links. In doing this, replace uses of
MediaWikiServices with dependency injection.

This hook handler file is similar to AbuseFilter's ToolLinksHandler
hook handler file but provides code for different hooks.

The tests for Hooks::onUserToolLinksEdit are moved to a test
file for the newly created ToolLinksHandler class.

No functionality should change in the user interface of CheckUser.

QA steps:
* Login to an account that has the checkuser group
* Load Special:Contributions for an account and check to see if
the subtitle links have links to check the current user
* Load Special:CheckUser and run a check using the 'get edits'
type for a user. Ensure there is a "check" link shown in the
toollinks for the users in the results.
* (Optional if above done) Load Special:CheckUserLog and ensure
that log items have a "checks on" toollink for the user that was
checked.

Bug: T328334
Change-Id: I82891cfb625c0c2222f5c2cc516b6b3287a7c824