Move Hooks::onSpecialPage_initList to a hook handler file
Move Hooks::onSpecialPage_initList to a hook handler file created
for just this hook. While doing this, the access to the config
'CheckUserEnableSpecialInvestigate' via a global variable is
replaced with a dependency injected config. By injecting the
config, the tests can be converted to unit tests that mock the
config value.

Bug: T328334
Bug: T315648
Change-Id: I0d08b6b674abaae63570fc49b1dc5fb4345dafb1