Move CheckUser services in the root folder to a new sub-folder
Move the CheckUser services that are in the "src" folder into
a "Services" folder. This is done to better group the services
that are used across the CheckUser extension. A change that is
dependent on this will add a new service that helps split up
Hooks.php.

QA steps:
* Ensure tests pass, especially the ServiceWiring.php tests
* Make sure no exceptions occur on CheckUser, CheckUser API or
Investigate

Bug: T328334
Change-Id: I8bffcc08153a22f1a22017d699eb0eee488aa427