Add cul_result_id and cul_result_plaintext_id to cu_log
Add the columns cul_result_id and cul_result_plaintext_id to the
cu_log table to store comment IDs in the comment table. The comment
IDs will refer to the result/outcome text left by the checkuser
which ran the check. This is optional and would be added after
the log entry is created.

A CU would be able to specify a result for the check within a
reasonable timeframe after the check was run. Once it is set it
would be final.

The idea is to allow CUs to give basic information about what the
check turned up. For example, it could indicate that socking was
detected but the CU would come back later to take another look or
even that a range has too many results for a check to be useful.

Bug: T190666
Change-Id: I021935f5a0df0283fef85983b5b1834df48890a0