Define SpecialInvestigateBlock::doesWrites as true for profiler
Define SpecialInvestigateBlock::doesWrites to return true. This
is because Special:InvestigateBlock will write to the DB when
submitting the form (to place a block, place user tags and
log blocks to Special:Log).

This is done to reduce logspam as the TransactionProfiler
indicates a potential issue every time the form is submitted.

Change-Id: Ia6d47c918a7c73f4de9a2ca11faa4f75ce8b97fd