Use separate user for selenium tests that require checkuser group
In a similar way to the CodeMirror extension a new user is created
for tests that need the test user to have the "checkuser" group.

This new test user has to be created by the Admin user as only the
Admin user can grant the "checkuser" group in the standard selenium
testing environment. To make the process quicker, the assigning of
the user group is done via the API instead of loading
Special:UserRights and submitting the form.

These changes make it possible to run the three selenium specs
concurrently as a separate user with the "checkuser" group is
created for each spec.

Bug: T340823
Change-Id: Ia4f64f1f56d1498fde3e197228f5672454d5865a