CheckUserLog API: Search by for users by target ID
To address slow read queries on enwiki, the CheckUserLog API has
been switched to use the same WHERE conditions for the target as
used for Special:CheckUserLog. This means that searching for a
username will use the ID of the user and limit the searched logs
to ones that can be made when the target is a user.

To do this, usernames which are not valid users will no longer
be accepted as valid targets and instead return an API error.
Other than this no user-facing change to the API has been made,
including what results would be returned if the target is a
user that exists.

This change means that the "cul_type_target" index on the cu_log
table can be used when searching for users using the API.
This should greatly increase the speed of the CheckUserLog API.

Bug: T341297
Change-Id: Ic0bf04e27f3ea0bb95d99f34aedca02db4eb46fd