Remove tables.sh and add version for addition of clienthints tables
Indicate that the client hints tables were added in MediaWiki
version 1.41. This follows the format of previous versions where
a comment is added just before the first database change made in
the given version inside SechamChangesHander.php

Also remove the tables.sh script as this is only applicable for
generating the SQL files for the two tables that have now been
added to the extension. As such, changes to these tables would
be made using the standard process without the need for this
bash script.

Bug: T258105
Follow-Up: I96f820ceb93f5e081f271e4dcc2be11309cedb35
Change-Id: I810be81887077678bc181ef09772f740270e4c6a