Fix test namespaces
Use "Tests" consistently to avoid maintaining two different versions,
update classes in the "Test" namespace.

Also add separate autoloader entries for Unit and Integration. PSR-4
requires the case of the namespace and the directory to match, but our
convention is to use lowercase "unit" and "integration" for directories.
This leads to an issue where if you try to run the PurgeOldDataTest (as
an example of test that uses a test trait) on its own, it will fail
because it can't find the class. For some reason, it does work when
running all tests, but still it needs to be fixed.

Change-Id: Ic9583740ec88bb05ef319f6a455fb34996c31859