clienthints: Return public explanation of the use of client hints
Return on a successful client hints REST API response, instead of
the boolean "true", a translatable message that indicates what the
use of client hints is.

This message is kept deliberately small to prevent too much data
transfer and a page on mediawiki.org is linked to that provides a
more detailed explanation as to why client hints data is collected.

Bug: T341093
Change-Id: Ia11141813e7d26c6180d8ff8d6f6917bf13eeda0