Use ConvertibleTimestamp::time instead of in-built time for prune
Use ConvertibleTimestamp::time instead of the PHP in-built time
for both the PruneOldData.php maintenance script and the
checkuserPruneCheckUserDataJob.

This is done as the tests for these use ConvertibleTimestamp
::setFakeTime and the documentation for this method says that this
sets the fake time for calls to ConvertibleTimestamp.

As such, replacing the usage will likely fix the flaky-ness of the
tests because the test would set the fake time as the "current time"
as measured by the data provider. Before this change, if the tests
ran slow enough, the first test could have the actual time returned
by time() be more than 100 seconds after the the fixed time generated
when the datasets were created.

Bug: T342422
Change-Id: I84ae77ed7b808d3219b31e3308b9b6c2364e6306