Add GENDER support to checkuser-logout
Add GENDER support to checkuser-logout by providing the username
for the account that was just logged out of by this user.

This message will be removed soon from the main branch as part of
the event tables migration, however, this is a small fix so it
should be okay to make this even if this message is facing removal
soon.

Bug: T342486
Change-Id: I8c4efd09fc85f2ea86ae4ad98a6a3877ea84405c