Temp account revision API: Return missingparam on missing IDs
Return a missingparam 400 response to the user if the IDs parameter
provided is empty. The ParamValidator treats the empty array as
having met the PARAM_REQUIRED constraint and therefore returns
an empty array for use in TemporaryAccountRevisionHandler::getData.

This is done by throwing an exception (in the same way that the
ParamValidator does in the case of a missing parameter) when the
count of the IDs array is 0. Without this the REST API returns a
500 error, which is not the correct response when the client was
the one "at fault" for not providing revision IDs.

Bug: T331413
Change-Id: I760aea21bad25fa7fad5785df450a4f298cfcd4c