Allow IP addresses as usernames in CheckUserPrivateEventLogFormatter
As evidenced in the associated task, the username used for a
failed login attempt that is stored in the CheckUser tables can
be an IP address on WMF wikis.

As IP addresses are considered invalid users by the RIGOR_VALID
constraint, the CheckUserPrivateEventLogFormatter needs to pass
RIGOR_NONE when the user object is created using UserFactory::
newFromName.

If this is not done, IP addresses being used as the username for
a failed login attempt would cause an InvalidArgumentException
to be thrown.

Where possible, this migration avoids reducing or removing data
that is collected by CheckUser. While the display of this data
may be later determined a bug, the current interface displays
this and it is caused by a user submitting the login form.

What:
* Pass UserRigorOptions::RIGOR_NONE to UserFactory::newFromName when
  creating the user object for CheckUserPrivateEventLogFormatter.
* Add a missing comment to the log formatter that allows searching
  for the usage of the logentry definition in the source code.
  This comment was included for all other messages with added
  message parameters in the CheckUserPrivateEventLogFormatter class.

Bug: T342629
Change-Id: Ia0c078a9cc8521b783c4e797df701c5fd0e47c22