Make some of the non-static data providers static
Make the non-static data providers in all test classes, except from
CheckUserUnionSelectQueryBuilderTest, static by modifying them to
no longer need $this. For the tests in HooksTest.php, this is done
by creating the AuthenticationResponse in the test methods as some
of these required a AuthenticationRequest mock to be provided when
creating the AuthenticationResponse object.

Bug: T337159
Change-Id: If0b4d2b3ba10e5974a54d5627a78f9419d4d5bb6