Prevent saving Client Hints using REST API when action is too old
Do not save Client Hints data that is provided via the REST API
when the provided reference ID refers to an action that was
performed over wgCheckUserClientHintsRestApiCutoff seconds ago.

This is done to help reduce the ability for users to send fake
and/or modified Client Hints data for a revision which has had
no data already saved. This also reduces the chance of another
user on the same IP address being able to save fake Client Hints
data for an edit made using an IP address.

What:
* Add the config wgCheckUserClientHintsRestApiCutoff which is set by
  default to 30 minutes (1800 seconds). This config contains the
  number of seconds that a user is given to submit the REST API
  request.
* Return a 403 response with associated message (instead of storing
  Client Hints data) if a request to the Client Hints REST API
  provides an ID which was made over
  wgCheckUserClientHintsRestApiCutoff seconds ago.
* Add a unit test for this behaviour as well as mocking getTimestamp
  for the RevisionRecord object to support this new check in other
  unit tests.

Bug: T342134
Change-Id: I190fbd7879b65f779956b255b7d72eaa91d2ab0a