clienthints: Allow Client Hints API to accept partial data
Allow calls to the Client Hints REST API to provide a subset of the
requested Client Hints data.

This is done to prevent the need to customise the required Client
Hints for each version of the browser (which is the issue in the
linked task). This also allows the browser to intentionally send
less Client Hints data if the privacy budget prevents sending
every Client Hints data item that was requested.

If a particular user decides to modify the REST API call to only
send part of the Client Hints data, then the part that was
provided should be stored to allow CUs to see at least some of the
data. It is also unlikely that the browser would send more Client
Hints data at a future point, so the partial response should be
treated as final.

If a call to the API sends no data, then future calls with data
could allow client hints data storage. These future calls are
likely to contain modified Client Hints data, but the time window
to insert this will be small (because of the patch this depends on)
and some data is better than no data.

If the browser was to send less or no data, this in itself can
be used by CUs as evidence to link users as most users will not
modify the API request (in a similar way to why User-Agent string
spoofing already is used as a way to link users). No data being
send with a User-Agent string indicating a Chromium browser would
indicate that a user prevented the API request or modified it to
include no data. In both cases this indicates that the user
removed this data (which again can be used as a way to link
to two users that did this behaviour).

What:
* Remove ParamValidator::PARAM_REQUIRED statements for all client
  hint types in UserAgentClientHintsHandler::getBodyValidator

Bug: T342729
Change-Id: I3d9da9a318359d4dfe586ebd2ed03a5c6316965b