clienthints REST API: Perform matching user check with RAW audience
The Client Hints REST API gets the user associated with the
revision with the RevisionRecord::FOR_PUBLIC audience when checking
to see if the performer of the revision is the same as the user who
called the API. This means that when checking a revision that has
been deleted or had the performer revision deleted, the check will
fail as ::getUser will return null.

This change makes the call to ::getUser use RevisionRecord::RAW,
such that the check will properly compare the users in the case
that the revision is deleted and therefore still allow Client Hints
data to be recorded. In normal usage, this could occur when a page
was deleted just after an edit was made to the page.

There will no vector to leak the deleted username as the ID of the
performer of the API request is the only information shown on a
mismatch.

What:
* Modify UserAgentClientHintsHandler::run to pass RevisionRecord::
  RAW to the RevisionRecord::getUser() calls.

Bug: T342790
Change-Id: Iaa48a28ac85342d71da2ae29a8e31b03317f12ab