Use class covers annotation instead of per-method annotations
Use class covers annotation instead of per-method annotations for
the methods that are marked as covering the class marked as covered
in the documentation block at the top of the file.

This excludes @covers for methods in HooksTest.php which also
cover a specific method of HookRunner.php. This is done as
the HooksTest.php file does not cover the entirety of the
HookRunner.php file. The @covers annotations for methods in
Hooks.php for these test methods are kept to reduce confusion
as to whether these should be in the HookRunner.php test files.

This is done because PHPUnit documentation recommends using a
covers annotation for the entire test file instead of per-method.
Also this is to bring consistency to CheckUser tests as they use
a combination of per class covers annotations and per method
annotations. Finally, this reduces the length of the test files
which should make them easier to read.

Also, the class @covers annotation for SpecialCheckUserLogTest.php
is fixed as well as the import of the CheckUserLogService in that
file.

Change-Id: I63f9c99265a7d59901147e95247dfab33715be6d