Use IGNORE option when inserting into cu_useragent_clienthints
Use the IGNORE option when inserting into cu_useragent_clienthints
so that any rows which already have the same name and value
combination do not cause an unique constraint violation when
attempting to insert.

This can happen when a replica DB is lagged and it read from to
determine whether there exists all the cu_useragent_clienthint
rows that would be used by the mapping rows. If the lag occurs,
then the code will think there needs to be new rows added to
cu_useragent_clienthints that are present on the primary.

Passing the IGNORE option on insert to the primary DB will
prevent this situation from causing a DB error because some
rows violate the unique constraint imposed by the index.

What:
* Pass 'IGNORE' as an option when inserting to the table
  cu_useragent_clienthints in UserAgentClientHintsManager::
  insertClientHintValues.

Bug: T342826
Change-Id: I6999bc5abaf2bb138f649e21245eb50adbb26d80