Use UserIdentity::LOCAL in PreliminaryCheckService when appropriate
Use UserIdentity::LOCAL in PreliminaryCheckService as the wiki
argument for the UserIdentityValue object that is passed to
UserGroupsManager::getUserGroups.

This is done as if the wikiId provided to PreliminaryCheckService
::getAdditionalLocalData is the wiki ID of the local wiki
(i.e. the argument is enwiki when the local wiki is enwiki),
providing this as the wikiId when creating the UserIdentityValue
will cause the check to fail as ::LOCAL (which is false) is not
equal to the wikiID (which is always a string).

Bug: T344403
Change-Id: I4d42cafa2bc0a83a10cf80caf12e655cfc82b846