clienthints: Remove server-side check for browser support
Remove the server-side check for Client Hints support in
::onSpecialPageBeforeExecute and ::onBeforePageDisplay.
This is done as edge caching would mean the ResourceLoader
module gets loaded or not loaded inconsistently depending
on the browser that caused the cache refresh.

Removing the check means that the ResourceLoader module
is always loaded for all browsers when all other checks
pass (i.e. whether Client Hints collection is enabled).
This means edge caches will have ResourceLoader module
for Client Hints provided if the page would need the
module if the browser supports Client Hints.

If the module is loaded for a browser with no Client
Hints API support, the JS code checks for this and
stops execution of the JS file before it tries to
use the API. This means for browsers with no
support, no behaviour will change.

Bug: T344679
Change-Id: If7ad5df2ece895528a81f46dde22d9d1ce648dbc