clienthints: Remove duplicate entries when converting to DB rows
When converting a ClientHintsData object to the equivalent database
rows, remove duplicate entries by using array_unique on the uach_value
column for names that accept an array (e.g. brands or fullVersionList).

This is done by converting all the items in the array to the string
form and then calling array_unique on the items in ClientHintsData
::toDatabaseRows. This ensures that the rows do not contain any
duplicates, as the cu_useragent_clienthints table does not support
including duplicate entries.

This fixes the issue with the insert to cu_useragent_clienthints_map
as the insert query will be generated with unique Client Hints data
rows. This is done instead of de-duplication in UserAgentClientHints
Manager::insertMappingRows as other methods would not be wrong in
assuming that the rows returned from ClientHintsData::toDatabaseRows
would be fine to insert to the DB as is.

A checkuser running a check is very likely to not find use in knowing
that duplicate entries were submitted, so de-duplication at the time
of insert is fine. To support storing that a item was duplicated,
a database schema change would be needed.

Bug: T344787
Change-Id: If6c5408f27ca22c8d228a1e61a5f2f50f5a985a1
(cherry picked from commit 92729522b4fdd2dbaee9823deeac3ffe94a80393)