selenium: run tests concurrently
Run the specs concurrently after fixing the code to support
concurrent execution of the test specs.

This is done by creating a new wdio.pre-specs.conf.js file that runs
one spec file in the pre-specs folder that creates an account with
the username "CheckUserAccount" and gives it the checkuser group.
This is done so that this runs first and completes before concurrent
execution of the other specs occurs.

These specs are run by a wrapper file that uses a method to run
the specs in an arbitrary environment detailed by the documentation
(https://v6.webdriver.io/docs/clioptions.html#run-the-test-runner-
programmatically). This wrapper file generates a random username and
password for the account that is given CheckUser rights, and passes
it the specs via a overridden config when using the Launcher.

Doing this via the method suggested in the task, according to
my understanding of the documentation, would not allow the specs
to be run concurrently as they would be run one at a time on one
worker.

This method increases the speed of the selenium tests by around
2 to 3 times.

Bug: T226869
Bug: T340823
Change-Id: I1e594a802caadb2f2eb048e6a6ae7efa51320eb9