clienthints: Return early on no DB rows in ::insertClientHintValues
Return early from UserAgentClientHintsManager::insertClientHintValues
with a good status if there is no data in the provided ClientHintsData
object.

This is done by checking if the ClientHintsData::toDatabaseRows
call in UserAgentClientHintsManager::insertClientHintValues returns
an empty array of rows. If this occurs, then the method returns
early with the good status.

No data to insert is not considered a "bad" thing that should cause
a 400 error as the browser may decide to send no Client Hints data
via the API or HTTP headers. In this case, there would be nothing
to store by CheckUser.

Bug: T344873
Change-Id: I98ca0a7b615acefca9c5130953ec291956cd99cb