SECURITY: Limit number of rows returned by ::toDatabaseRows
In the ClientHintsData value object, the "brands" and
"fullVersion List" properties are defined as arrays. Each
array item stores a brand name and then version number
(e.g. Chrome 114). These arrays usually have around 2 to
4 items per array. However, there is currently no limit
on the number of items these arrays can hold.

The ClientHintsData::toDatabaseRows method returns the
Client Hints data as database rows. These returned rows
are inserted into the cu_useragent_clienthints table.
The ::toDatabaseRows method returns a unique row for
each item in the arrays. If the request that sent the
data is maliciously modified, thousands of rows could
be inserted at once.

This patch limits the number of database rows returned
that are from the brands and fullVersionList arrays to
10 each. This means that a maximum of 29 rows can be
returned from ::toDatabaseRows (10 brands, 10
fullVersionList, and 9 from the other properties that
map 1 property to 1 row in the conversion).

While display of this data is not implemented yet, storing more
than 10 items from either of these arrays in the database is
unneccessary as there will not be the space to display them.

Bug: T344923
Change-Id: Ifb46cc364bf9c1a256f876aa0a2f9bdb7c3773c8