clienthints: Only call implode when the item is an array
Add code to check if the item is an array before attempting to
implode it in ::toDatabaseRows. If the item is a string or is
numeric, convert the item to a string.

While the expected format of the brands and fullVersionList
are arrays, accepting strings and numbers is deemed okay as
this may still contain useful data (such as a pre-imploded
brand and version number, or just a version number). As the
Client Hints API is marked as experimental by MDN web docs,
the format could still change to not be this structure.

If the type is not supported, the data is just ignored by
::toDatabaseRows and therefore no associated row is returned.
Booleans in this context do not make sense, so are deliberately
ignored. Null is treated as no value, so is also ignored. Other
types do not make sense in this context.

Bug: T344993
Change-Id: I348bfcd2abe1726b99d48d90d135f9d79c7c2e91