Order the "brand" and "version" of brands and fullVersionList array
Insert the correct uach_value in the databse, irrespective of the order of "brand" and "version".
Alphabetically sort the brands and fullVersionList arrays.
After sorting, "brand" should always appear before the "version".
Add Tests in ClientHintsDataTest.php to test the new change.

Rectified the new code style as per the MediaWiki standard
Added Test case to test the new code with the wrong order of "version" and "brand" of brands and fullVersionList array.

Change in Unit Test data as per the post review comments

Bug: T344979
Change-Id: I1d2cf1c69b5e52d78240a30398a6972fdcc137f4