clienthints: Use IGNORE when inserting mapping rows
Why:
* A user may submit multiple Client Hints API requests at once when
  the replica DB is lagged.
* This causes a DB error as the lagged replica DBs cannot see the
  changes on the primary DB.

What:
* Call the ::ignore builder method when constructing the insert
  query to insert mapping rows.

The use of the ::ignore method makes the SQL statement an
"INSERT IGNORE" which ignores invalid rows instead of raising a DB
error. A row can be invalid when it would violate the unique index
constraint (which in this case happens when the map row already
exists). The other situations where a row could be "invalid" should
already be guarded against or are not possible, so making the query
ignore invalid rows shouldn't hide other existing bugs.

Bug: T345144
Change-Id: I3e3c80cc2372515fa40e1b51174d23332a03ecaf