Store event in cu_private_event when related log entry cannot be found
In some rare cases, code that created the related DatabaseLogEntry ended up
with database rollback to unknown circumstances. In such scenarios hooks
failed to update user data due to failure in fetching DatabaseLogEntry - the
id was there, but entry couldn't be found in DB.

In such situations, lets store data in cu_private_event so information is not
lost.

Bug: T343983
Change-Id: I30c2dba1a9b29b660058b56f393542fc8ad4cf04