In SpecialCheckUserTest don't have the test sysop block themselves
testDoMassUserBlockInternal did three blocks with the performer set to
$this->getTestSysop(): the second being a self-block and the third being
an IP block.

The proposed new block cache is not quite as reliable as User::$mBlock,
so after the core change is merged, BlockPermissionChecker notices that
the performer is blocked and denies the third block.

I don't think it was intentional to test blocking with a blocked
performer.

So, use a different user as the target.

Needed-By: I25b84b0a8f9cacd0908a415b3a4a50ff7ecc72f4
Bug: T345683
Change-Id: If57d4e910d35f386028afd9cb900d78f3b6a0e13