Use real users in CheckUserPrivateEventLogFormatterTest
LogFormatterTestCase::doTestLogFormatter() overrides the UserFactory
with one that always returns user objects with user_id=42.
CheckUserPrivateEventLogFormatter asks this user for block details,
which previously worked because user blocks were fetched by name.
However, for T346683, I am trying to load them by ID instead.

So, do a proper integration test, without overriding services. Use real
users.

Most of doTestLogFormatter() is devoted to installing mocks, so it's
simplest to duplicate the parts of it that are not mock-related.

testLogDatabaseRowsForHiddenUserAndAuthorityHasSuppressGroup() was a
near-duplicate so I factored out a common test and used a provider.

Bug: T346683
Needed-By: I408acf7a57b0100fe18c455fc13141277a598925
Change-Id: If78b66a3da3b7d03c8448363995ecf3bdc211ab4