MoveLogEntriesFromCuChangesTest: Set fake time
MoveLogEntriesFromCuChangesTest depends on the current time
by calling ConvertibleTimestamp::now(). From time to time, the tests
don't execute at the expected second and fail. A recent example
from the GrowthExperiments extension is available as
https://integration.wikimedia.org/ci/job/wmf-quibble-vendor-mysql-php74-docker/44179/console

To prevent this from happening, set a fake time via
ConvertibleTimestamp::setFakeTime. Since the whole test depends
on time, do it in the setUp function.

Change-Id: Ie6824149fa8fc29fe08cb9a976e9d95742647b3c