Turn TemplateParserMockTest into an actual mock class
Classes ending in `Test` are considered test classes by PHPUnit, hence
it expects to find tests inside. But TemplateParserMockTest is a mock
implementation of TemplateParser, not a test, hence it should not be
disguised as a test class. So, rename it to MockTemplateParser,
following the convention used in core's tests/phpunit/mocks.

Also put the class in the right namespace according to PSR4
autoloading, so that it can be loaded properly while running tests: this
class is referenced in other test classes, and therefore it needs to be
loaded for those references to work when said tests are run individually
(or when using paratest, T50217).

Change-Id: Ib6eee952d69bfad6e51b7aa277b663a50fa874af