Add grant configuration
Add a 'checkuser' grant for full checkuser actions, and a
'checkuser-temporary-account' grant for checkuser accounts on
temp users.

Previously some of this configuration was in Wikimedia-specific
repos (operations/mediawiki-config and WikimediaMessages, see
Ibd6aab68379 and I6fde899a97). Note that previously there was no way
to grant the 'checkuser' permission.

Bug: T290790
Bug: T359537
Change-Id: I6fde899a97e9fdad20b597401c7860861660df58