Abstract parameter validation from AbstractTemporaryAccountHandler
AbstractTemporaryAccountHandler was built with the assumption that
all endpoints would require a name parameter. With the introduction
of "/checkuser/v0/temporaryaccount/ips/{ip}", this is no longer true
as this endpoint requires an IP instead. It's still useful to build
off of the main abstract, as it contains user permissions validations
and logging but needs to be able to accomodate both the name and ip
parameters.

- add AbstractTemporaryAccountNameHandler and split name validation
  functions into this class
- extend this new class from any endpoint that uses the name parameter

Bug: T354542
Change-Id: I782cacdb107c6a72f5edfe11871177fec9101db8