Prevent IP reveal buttons from being added to Special:Log
In most cases, the temporary user links on Special:Log do not
have corresponding IP addresses that can be looked up using the log
ID in the HTML.

This leads to a confusing situation where pressing a reveal button
displays no IP, or an IP that is unrelated to the log line.

It is better not to add the buttons that to introduce this
confusion.

Bug: T370083
Change-Id: I30fd96f3a8a568eba1b0f26a2b5803191fc74378