Enable temporary accounts in two tests
These tests seem to fail when temporary accounts are disabled (e.g.
locally or, as with Ibd020a5260, temporarily in another extension’s CI).
I don’t know if there might be better ways to fix them, but
TempUserTestTrait is already used in this test class, so IMHO it’s
reasonable to just call enableAutoCreateTempUser() in these tests.

Change-Id: I69746301c5cd9c8908271a40f8083e039beca6d2