Scope every use of `enableAutoCreateTempUser` in test suite
We continue to see interactions between the CheckUser TempUser tests
and other extensions (specifically Math and Translate). To try and
eliminate these and all future such issues, try adding a scope to
all invocations of `enableAutoCreateTempUser`.

Create a CheckUser-specific `TempUserTestTrait` that adds the
`genPattern` argument during `enable`, and the `matchPattern` during
`disable`.

Bug: T377197
Change-Id: I9de6445a424d26c2fe2f957ff66111bd9374d848