Hide username warnings while usernames are unsupported by Special:GlobalContributions
Usernames are currently not a supported input on Special:GlobalContributions and
attempting to search using one will result in an invalid input error.
Independent of this, Special:Contributions and children classes will output username
errors if a username is unregistered alongside Special:GC's invalid input error.

As this could cause confusion as to whether or not a valid username input is allowed
(it's not), hide the error until Special:GC supports usernames.

- Return early in SpecialGlobalContributions->contributionsSub if a username is the
  target, as it's an invalid input and shouldn't need any of the links generated

Bug: T377002
Change-Id: I7fa8d186d12ae3b59c0b8403621423ed29628844