Replace deprecated ConvertibleTimestamp::microtime with hrtime
Follows-up I838d33c5e9 (CheckUser), and Ie48d9400cf (libs/Timestamp)
which deprecates the newly-added microtime in favour of hrtime.

While at it, remove the EventLogger in favour of direct calls.
I thought maybe this method existed to ease mocking in tests, but
I found no PHPUnit test mocking the return value of this method.

Instead, the built-in setFakeTime() of ConvertibleTimestamp is used,
which indeed exists for that purpose so that it can be used directly
in code.

The code measuring latency is SpecialInvestigate and not EventLogger,
so I've placed it there instead.

Bug: T245464
Change-Id: Ia6297d34f62fe3cddaa8cac048d6dec0f46ada80