Show an error when an API lookup fails on Special:GlobalContributions
Why:
Special:GC makes calls out to external wikis to check for permissions
before pulling in revisions. If this lookup fails, Special:GC has no way
of knowing whether or not the user has the right to view the requested
revisions. Show an error to the user in this stage to this effect rather
than an error that assumes the user doesn't have the right permissions.

What:
- Add an API lookup error message whenever CheckUserApiRequestAggregator
  doesn't return a result as expected

Bug: T379107
Change-Id: If14c024634f31d6a96b90855a18736d06488bf11