Update SpecialIPContributions->getTargetField to use new exclusion fields
Why:
At the time the page was created, the `excludenamed` and `excludetemp` didn't
exist. As of e5667bafc18bd203366fae7af3a0c0954786956a, these properites are
used to stop unnecessary API calls on IP-only inputs and as such, this field
should be updated to take advantage of this.

What:
- add `excludenamed` and `excludetemp` properties to Special:IPContributions'
  input field

Bug: T380407
Change-Id: I02b6fc50991f35a67d5c69a144b90e47a1dab802