Updates for multiblocks
Passing true to BlockUser::placeBlock() is deprecated and throws an
exception if the user is blocked multiple times. So, on
Special:InvestigateBlock, when multiblocks is enabled, instead of a
reblock checkbox, have an "add block" checkbox. With multiblocks,
reblocking is awkward but adding blocks is easy and should meet the
use case for reblocking.

Also:

* In PerformRetroactiveAutoblockHandlerTest, avoid passing
  UserIdentity to Block::setTarget(), since this will soon be
  deprecated. Use the new factory method DatabaseBlockStore::
  newUnsaved().
* Avoid deprecated method BlockPermissionCheckerFactory::
  newBlockPermissionChecker().

Bug: T387723
Bug: T385966
Change-Id: I41f520afd942d5a0006d24c4feab5db5b5a44126