styles: Harden design token usage in calculations
As design token values could be numeric values with
units, but also upcoming CSS custom properties,
it's useful to rely on CSS `calc()` function to
ensure that the values are correctly calculated.
This will also bring us closer to removing Less parser
in future and rely on CSS standard only.

Bug: T389234
Change-Id: Iae5a970b71faf523d298cddd6b4b6c89cae6fe54