Reduce selects in UserAgentClientHintsManager when check existence
The manager has to check if the client hint data already have an id in
the `cu_useragent_clienthints` table.
This is done via a select per client hint.
After the insert of missing data the data are read again to get the ids.
Both selects are done as single-row select per client hint data.

Combine the select into one multi-row select.

Change-Id: I880172a9d2fc205a044fb6481e4b8c59afc4f0a2