tests: Make HTML class check assertion more robust
SpecialGlobalContributionsTest needs to assert that two classes are
present on an element. Previously, it was just checking the HTML for
presence of a substring with both classes listed consecutively in a
given order. This is prone to breaking if there are changes to the HTML
structure. So, use a regexp that accounts for flipped ordering, as well
as more classes being added between the two we're checking. Also add
word boundaries to ensure we don't match unrelated classes; this is
particularly useful for the `external` class, as "external" could be
part of an unrelated class.

Bug: T393178
Change-Id: Iabbfe5d44bfb6d44b894df79c415b44265485e36