SUL3: Always set client hint headers on auth shared domain
This setting controls whether to send the Accept-CH headers on all
auth shared domain or not. When set to true, all requests on the
shared domain send the `Accept-CH` header to collect client hints
data for that client.

Test plan
=========

With this patch checked out, verify that setting the new config's
value to `true` on the auth shared domain does the following:

1. For every request, sends response header `Accept-CH` with
   appropriate values e.g.

```
accept-ch Sec-CH-UA, Sec-CH-UA-Arch, Sec-CH-UA-Bitness, Sec-CH-UA-Form-Factor,
           Sec-CH-UA-Full-Version-List, Sec-CH-UA-Mobile, Sec-CH-UA-Model,
           Sec-CH-UA-Platform, Sec-CH-UA-Platform-Version, Sec-CH-UA-WoW64`
```

2. On any other domain that is not the auth shared domain, an
   empty `Accept-CH` response header is sent.

The test will be done when deploying the config patch linked to
this.

Bug: T395185
Change-Id: I24ebb8e8435bc30ac528a107b0be5d430c01e3a6