UserInfoCard: prevent default link behavior with "click"
Currently, when a user clicks on the user info button (which is realy
a link under the hood) to display a User Info Card, a "#" character
gets appended to the URL for the page and the scroll position is
reset to the top, since the browser thinks it is trying to navigate
to a non-existent anchor link.

To fix this, the "mousedown" event name has been replaced with "click";
click handles both mouse button press as well as mouse button release,
(and should also cover touch events). If the goal is preventing the
default link behavior whenever JS is running, "click" is more likely
to cover everything.

Bug: T398323
Change-Id: I5e41b607370ba5e6d5653bcd13cfb98639773565