Reset time limit (max_execution_time) in SpecialCheckUser tests
SpecialCheckUser::execute(), in some circumstances, sets the PHP
execution time limit to 60 seconds (and has done so since 2008); this is
presumably meant to be an increase from whatever the default is, but it
can also end up imposing a limit that previously did not exist. The unit
tests should run without any time limit, so add a PHPUnit after-test
hook to reset the time limit again.

Bug: T402390
Change-Id: Ia455bb7b70af95208fa8a3f420279202d53f1c24